+++ components/cc2500/cc2500.h
@@ -251,7 +251,7 @@ public:
 
 	void set_parent(CC2500Component *parent);
 	void send(uint8_t *data, uint8_t length);
-	virtual bool receive(uint8_t *data, uint8_t length, uint8_t rssi, uint8_t lqi) {
+	virtual bool receive(uint8_t *data, uint8_t length, float rssi, uint8_t lqi) {
 		return this->receive(data, length);
 	}
 	virtual bool receive(uint8_t *data, uint8_t length) {
